backend(python)/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-wireguard
